[irep2] consolidate template .cpp files into irep2_utils.cpp
Folds src/irep2/irep2_templates.cpp and src/irep2/irep2_template_utils.cpp
into src/irep2/irep2_utils.cpp so the field-dispatch overload catalogue
(type_to_string / do_type_lt / do_type_crc / do_type_hash / do_get_sub_expr
/ do_count_sub_exprs / call_expr_delegate / call_type_delegate overloads)
ships from one translation unit instead of three. The two headers
(irep2_templates.h, irep2_template_utils.h) keep their declarations since
they are still consumed by irep2_dispatch.h and the unit tests.

Move the expr_id / type_id "fits in a byte" static_asserts next to their
respective name tables in irep2_expr.cpp and irep2_type.cpp where the
invariant they guard lives.
