+++ src/irep2/CMakeLists.txt
@@ -1,8 +1,6 @@
 add_library(irep2
   irep2_type.cpp
   irep2_expr.cpp
-  irep2_templates.cpp
-  irep2_template_utils.cpp
   irep2_utils.cpp
 )
